Rename files and fix bug
#79
