+++ frontend/src/components/features/dashboard/core/DishesList.jsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { Container } from 'react-bootstrap';
-import Scrollable from './../../../common/Scrollable';
-import RestaurantCard from './RestaurantCard';
+import Scrollable from '../../../common/Scrollable';
+import DishCard from './DishCard';
 
 const DishesList = ({ dishes, setSelectedDish }) => {
 	return <Container className="px-4">
 		<Scrollable height={425}>
 			{dishes.map(restaurant => (
-				<RestaurantCard
+				<DishCard
 					key={restaurant.id}
 					dish={restaurant}
 					setSelectedDish={setSelectedDish}
